+++ src/main/java/com/atlauncher/data/Settings.java
@@ -95,6 +95,7 @@ public class Settings {
     public int concurrentConnections = 8;
     public int connectionTimeout = 30;
     public boolean dontUseHttp2 = false;
+    public boolean dontValidateModpacksChDownloads = false;
     public boolean enableProxy = false;
     public String proxyHost = "";
     public int proxyPort = 8080;